+++ src/main/webapp/memberlist_search.jsp
@@ -32,11 +32,12 @@
 							<option value = "1">항목</option>
 							<option value = "member_id"<% if(opt.equals("member_id")){%> selected<%}%>>ID</option>
 							<option value = "member_name"<% if(opt.equals("member_name")){%> selected<%}%>>이름</option>
+							<option value = "dept_Name"<% if(opt.equals("dept_Name")){%> selected<%}%>>부서명</option>
 							<option value = "zipcode"<% if(opt.equals("zipcode")){%> selected<%}%>>우편번호</option>
 							<option value = "lot_addr"<% if(opt.equals("lot_addr")){%> selected<%}%>>지번주소</option>
 							<option value = "road_addr"<% if(opt.equals("road_addr")){%> selected<%}%>>도로명주소</option>
 						</select>
-						<input type = "text" name = "search" value = <%=search %>>
+						<input type = "search" name = "search" value = <%=search %>>
 						<input type = "submit"  id = "submit" value = "검색">
 					</td>
 				</tr>
@@ -69,9 +70,10 @@
 			</div>
 			<table border = "1" id = "list">
 			<tr style = "background: #7D9D9C; color: white;">
-				<td>num</td>
+				<td>No</td>
 				<td>ID</td>
 				<td>이름</td>
+				<td>부서명</td>
 				<td>성별</td>
 				<td>생년월일</td>
 				<td>HP</td>
@@ -92,11 +94,16 @@
 				PreparedStatement pstmt = null;
 				ResultSet rs = null;
 				
-				String sql = "SELECT * FROM shopping_member WHERE 1 = 1 ";
+				String sql = "SELECT * FROM shopping_member as s INNER JOIN dept as d ON s.dept_No = d.dept_No WHERE 1 = 1 ";
 				
 				if(!opt.equals("1")){
-					sql += "and " + opt + " like '%" + search + "%'";
+					if(opt.equals("dept_Name")) {
+						sql += "and " + opt + " like '%" + search + "%'";
+					} else {
+						sql += "and " + opt + " like '%" + search + "%'";
+					}
 				}
+				
 				
 				if(male.equals("male") && female.equals("female")){
 					sql += "and (member_gender = '" + male + "'" +  "or member_gender = '" + female + "')";
@@ -124,7 +131,8 @@
 				}else{
 				}
 				
-				sql += "ORDER BY member_id ASC";
+				sql += " ORDER BY member_id ASC";
+				
 				pstmt = conn.prepareStatement(sql);
 				rs = pstmt.executeQuery();
 				
@@ -133,6 +141,7 @@
 					count++;
 					String member_id = rs.getString("member_id");
 					String name = rs.getString("member_name");
+					String dept = rs.getString("dept_Name");
 					String gender = rs.getString("member_gender");
 					String birth_y = rs.getString("member_birth_y");
 					String birth_m = rs.getString("member_birth_m");
@@ -156,6 +165,7 @@
 				<td><%=i%></td>
 				<td><%=member_id%></td>
 				<td><%=name %></td>
+				<td><%=dept %></td>
 				<td><%=gender %></td>
 				<td><%=birth_y%>/<%=birth_m%>/<%=birth_d%>(<%=birth_gn%>)</td>
 				<td><%=HP1%>-<%=HP2%>-<%=HP3%></td>
@@ -183,7 +193,15 @@
 <script>
 	function mem_delete(id) {
 		if (confirm(id + "님의 계정을 삭제하시겠습니까?") == true) {
-			location.href="memberlist_delete.jsp?target=" + id;
+			var manage_num = prompt("관리자 비밀번호를 입력해주세요.");
+	        if(manage_num == "0000") {
+				location.href="memberlist_delete.jsp?target=" + id;
+	        } else if(manage_num == null){
+				return false;
+	        } else {
+	        	alert("관리자비밀번호가 일치하지 않습니다.");
+				return false;
+			}
 		} else {
 			return false;
 		}
